+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Electric and Coulomb constants added. [#47](https://github.com/itt-ustutt/quantity/issues/47)
+- ATM and POISE added as derived units. [#46](https://github.com/itt-ustutt/quantity/issues/46)
 
 ## [0.13.0] - 2026-01-06
 ### Changed